move `event_bus` code to a single file
also remove unneeded `Metric` event
