+++ bottlecap/src/lib.rs
@@ -19,7 +19,6 @@
 
 pub mod config;
 pub mod event_bus;
-pub mod events;
 pub mod fips;
 pub mod http;
 pub mod lifecycle;
