+++ bottlecap/src/logs/agent.rs
@@ -1,7 +1,7 @@
 use std::sync::{Arc, Mutex};
 use tokio::sync::mpsc::{self, Sender};
 
-use crate::events::Event;
+use crate::event_bus::Event;
 use crate::logs::{aggregator::Aggregator, processor::LogsProcessor};
 use crate::tags;
 use crate::telemetry::events::TelemetryEvent;
